Move infotagger import to compatibility module
