+++ resources/lib/youtube_plugin/kodion/compatibility/__init__.py
@@ -26,6 +26,8 @@ try:
     import xbmcplugin
     import xbmcvfs
 
+    from infotagger.listitem import set_info_tag
+
 except ImportError:
     import BaseHTTPServer
     from contextlib import contextmanager as _contextmanager
@@ -91,12 +93,36 @@ except ImportError:
     xbmcvfs.File = _file_closer
     xbmcvfs.translatePath = xbmc.translatePath
 
+    def set_info_tag(listitem, infolabels, tag_type, *_args, **_kwargs):
+        listitem.setInfo(tag_type, infolabels)
+        return ListItemInfoTag(listitem, tag_type)
+
+
+    class ListItemInfoTag(object):
+        __slots__ = ('__li__',)
+
+        def __init__(self, listitem, *_args, **_kwargs):
+            self.__li__ = listitem
+
+        def add_stream_info(self, *args, **kwargs):
+            return self.__li__.addStreamInfo(*args, **kwargs)
+
+        def set_resume_point(self,
+                             infoproperties,
+                             resume_key='ResumeTime',
+                             total_key='TotalTime'):
+            if resume_key in infoproperties:
+                infoproperties[resume_key] = str(infoproperties[resume_key])
+            if total_key in infoproperties:
+                infoproperties[total_key] = str(infoproperties[total_key])
+
 
 __all__ = (
     'BaseHTTPServer',
     'parse_qs',
     'parse_qsl',
     'quote',
+    'set_info_tag',
     'unescape',
     'unquote',
     'urlencode',
