+++ resources/lib/youtube_plugin/kodion/ui/xbmc/xbmc_items.py
@@ -11,35 +11,9 @@
 from __future__ import absolute_import, division, unicode_literals
 
 from . import info_labels
-from ...compatibility import xbmcgui
+from ...compatibility import set_info_tag, xbmcgui
 from ...items import AudioItem, UriItem, VideoItem
 from ...utils import current_system_version, datetime_parser
-
-try:
-    from infotagger.listitem import set_info_tag
-except ImportError:
-    def set_info_tag(listitem, infolabels, tag_type, *_args, **_kwargs):
-        listitem.setInfo(tag_type, infolabels)
-        return ListItemInfoTag(listitem, tag_type)
-
-
-    class ListItemInfoTag(object):
-        __slots__ = ('__li__',)
-
-        def __init__(self, listitem, *_args, **_kwargs):
-            self.__li__ = listitem
-
-        def add_stream_info(self, *args, **kwargs):
-            return self.__li__.addStreamInfo(*args, **kwargs)
-
-        def set_resume_point(self,
-                             infoproperties,
-                             resume_key='ResumeTime',
-                             total_key='TotalTime'):
-            if resume_key in infoproperties:
-                infoproperties[resume_key] = str(infoproperties[resume_key])
-            if total_key in infoproperties:
-                infoproperties[total_key] = str(infoproperties[total_key])
 
 
 def video_playback_item(context, video_item, show_fanart=None):
